+++ packages/pyroscope-flamegraph/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.2.0...@pyroscope/flamegraph@0.2.1) (2022-03-23)
+
+
+### Bug Fixes
+
+* **flamegraph:** clicking on anywhere on a row selects that row ([#969](https://github.com/pyroscope-io/pyroscope/issues/969)) ([ee84788](https://github.com/pyroscope-io/pyroscope/commit/ee8478812743e1381818e769706df83506ed6f53))
+
+
+
+
+
 # [0.2.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.1.3...@pyroscope/flamegraph@0.2.0) (2022-03-15)