+++ packages/pyroscope-models/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.4](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/models@0.0.3...@pyroscope/models@0.0.4) (2022-03-23)
+
+**Note:** Version bump only for package @pyroscope/models
+
+
+
+
+
 ## [0.0.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/models@0.0.2...@pyroscope/models@0.0.3) (2022-03-14)
 
 **Note:** Version bump only for package @pyroscope/models